aider: Renamed the refresh button to 'Force Refresh'.
# Context:
USER: rename the refresh button to 'Force Refresh'
ASSISTANT: To rename the "Refresh" button to "Force Refresh", we need to modify the `main.go` file. We will change the label of the menu item from "Refresh" to "Force Refresh".

Here are the changes:

```go
main.go
<<<<<<< ORIGINAL
	mRefresh := systray.AddMenuItem("Refresh", "Refresh the data")
=======
	mRefresh := systray.AddMenuItem("Force Refresh", "Force Refresh the data")
>>>>>>> UPDATED
```

This change will rename the "Refresh" button to "Force Refresh".