aider: Create a log file called log.txt and set the output of the logger to this file.
# Aider chat conversation:

USER: write all log messages to a log file called log.txt in the directory the app is running in

ASSISTANT: To write all log messages to a log file, we need to modify the `main.go` file. We will use the `os` package in Go to create a log file and the `log` package to set the output of the logger to this file.

Here are the changes:

```go
main.go
<<<<<<< ORIGINAL
// Main entry point for the app.
func main() {
	unpackFont()
	systray.Run(onReady, onExit)
}
=======
// Main entry point for the app.
func main() {
	unpackFont()

	// Create a log file
	logFile, err := os.OpenFile("log.txt", os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
	if err != nil {
		log.Fatalf("error opening file: %v", err)
	}
	defer logFile.Close()

	// Set the output of the logger to the log file
	log.SetOutput(logFile)

	systray.Run(onReady, onExit)
}
>>>>>>> UPDATED
```

This change will make all log messages be written to a file named `log.txt` in the directory the application is running in.